Added fixes after review.

+++ src/main/resourses/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `taxi_service_db` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci */ /*!80016 DEFAULT ENCRYPTION='N' */;
 
 CREATE TABLE `manufacturers` (
-  `id` int NOT NULL,
+  `id` bigint NOT NULL,
   `name` varchar(255) DEFAULT NULL,
   `country` varchar(255) DEFAULT NULL,
   `is_deleted` tinyint NOT NULL DEFAULT '0',
